refactor: change structure admin features
